Refactor building validations
You should use an array instead of a regex for checking the states. It
makes it easier to read if you store it in a constant instead of
directly in the validation.

You also want to make sure that when you do use regular expression that
you anchor your exression. Your previous regex for the city was only
checking that there was at least one character that matched. This new
one will check that all of the characters match, from the beginning of
the string to the end.